+++ Tests/Unit/MSFT_ADOrganizationalUnit.Tests.ps1
@@ -51,317 +51,627 @@ try
         $testAbsentParams = $testPresentParams.Clone()
         $testAbsentParams['Ensure'] = 'Absent'
 
-        $protectedFakeAdOu = @{
-            Name                            = $testPresentParams.Name
-            ProtectedFromAccidentalDeletion = $true
-            Description                     = $testPresentParams.Description
+        $mockName = 'TestOU'
+        $mockPath = 'OU=Fake,DC=contoso,DC=com'
+        $mockDistinguishedName = 'OU=' + $mockName + ',' + $mockPath
+
+        $mockResource = @{
+            Name                            = $mockName
+            Path                            = $mockPath
+            Description                     = 'Test AD OU description'
+            ProtectedFromAccidentalDeletion = $false
+            DistinguishedName               = $mockDistinguishedName
+            Ensure                          = 'Present'
         }
+
+        $mockAbsentResource = @{
+            Name                            = $mockResource.Name
+            Path                            = $mockResource.Path
+            Description                     = $null
+            ProtectedFromAccidentalDeletion = $null
+            DistinguishedName               = $null
+            Ensure                          = 'Absent'
+        }
+
+        $mockChangedResource = @{
+            Description                     = 'Changed Test AD OU description'
+            ProtectedFromAccidentalDeletion = $true
+        }
+
+        $mockProtectedAdOu = $mockResource.Clone()
+        $mockProtectedAdOu['ProtectedFromAccidentalDeletion'] = $true
+
+        $mockGetTargetResourcePresentResult = $mockResource.Clone()
+        $mockGetTargetResourcePresentResult['Ensure'] = 'Present'
+
+        $mockGetTargetResourceAbsentResult = $mockResource.Clone()
+        $mockGetTargetResourceAbsentResult['Ensure'] = 'Absent'
 
         #region Function Get-TargetResource
         Describe 'ADOrganizationalUnit\Get-TargetResource' {
-            It 'Returns a "System.Collections.Hashtable" object type' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource -is [System.Collections.Hashtable] | Should -BeTrue
-            }
-
-            It 'Returns "Ensure" = "Present" when OU exists' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource.Ensure | Should -Be 'Present'
-            }
-
-            It 'Returns "Ensure" = "Absent" when OU does not exist' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource.Ensure | Should -Be 'Absent'
-            }
-
-            It 'Returns "ProtectedFromAccidentalDeletion" = "$true" when OU is protected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource.ProtectedFromAccidentalDeletion | Should -BeTrue
-            }
-
-            It 'Returns "ProtectedFromAccidentalDeletion" = "$false" when OU is not protected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith {
-                    $unprotectedFakeAdOu = $protectedFakeAdOu.Clone()
-                    $unprotectedFakeAdOu['ProtectedFromAccidentalDeletion'] = $false
-                    return [PSCustomObject] $unprotectedFakeAdOu
-                }
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource.ProtectedFromAccidentalDeletion | Should -BeFalse
-            }
-
-            It 'Returns an empty description' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith {
-                    $noDescriptionFakeAdOu = $protectedFakeAdOu.Clone()
-                    $noDescriptionFakeAdOu['Description'] = ''
-                    return [PSCustomObject] $noDescriptionFakeAdOu
+            BeforeAll {
+                $getTargetResourceParams = @{
+                    Name = $mockResource.Name
+                    Path = $mockResource.Path
                 }
 
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-
-                $targetResource.Description | Should -BeNullOrEmpty
-            }
-
-            It 'Returns "Ensure" = "Absent" when OU parent path does not exist' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith {
-                    throw New-Object Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException
+                $mockGetADOrganizationUnitResult = @{
+                    Name                            = $mockResource.Name
+                    Path                            = $mockResource.Path
+                    Description                     = $mockResource.Description
+                    ProtectedFromAccidentalDeletion = $mockResource.ProtectedFromAccidentalDeletion
+                    DistinguishedName               = $mockResource.DistinguishedName
                 }
 
-                $targetResource = Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path
-                $targetResource.Ensure | Should -Be 'Absent'
+                Mock -CommandName Assert-Module
             }
 
-            It 'Should throw the correct error if an unknown error occurs' {
-                $errorMessage = 'Unknown Error'
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { throw $errorMessage }
+            Context 'When the resource is Present' {
+                BeforeAll {
+                    Mock -CommandName Get-ADOrganizationalUnit -MockWith { $mockGetADOrganizationUnitResult }
 
-                { Get-TargetResource -Name $testPresentParams.Name -Path $testPresentParams.Path } |
-                    Should -Throw $errorMessage
+                    $result = Get-TargetResource @getTargetResourceParams
+                }
+
+                foreach ($property in $mockResource.Keys)
+                {
+                    It "Should return the correct $property property" {
+                        $result.$property | Should -Be $mockResource.$property
+                    }
+                }
+
+                Context 'When the OU is protected' {
+                    BeforeAll {
+
+                        $mockGetADOrganizationUnitProtectedResult = $mockGetADOrganizationUnitResult.Clone()
+                        $mockGetADOrganizationUnitProtectedResult['ProtectedFromAccidentalDeletion'] = $true
+
+                        Mock -CommandName Get-ADOrganizationalUnit -MockWith { $mockGetADOrganizationUnitProtectedResult }
+                    }
+
+                    It 'Should return the desired result' {
+                        $targetResource = Get-TargetResource @getTargetResourceParams
+
+                        $targetResource.ProtectedFromAccidentalDeletion | Should -BeTrue
+                    }
+                }
+            }
+
+            Context 'When the resource is Absent' {
+                BeforeAll {
+                    Mock -CommandName Get-ADOrganizationalUnit
+
+                    $result = Get-TargetResource @getTargetResourceParams
+                }
+
+                foreach ($property in $mockAbsentResource.Keys)
+                {
+                    It "Should return the correct $property property" {
+                        $result.$property | Should -Be $mockAbsentResource.$property
+                    }
+                }
+            }
+
+            Context 'When the OU parent path does not exist' {
+                It 'Returns the correct result' {
+                    Mock -CommandName Get-ADOrganizationalUnit -MockWith {
+                        throw [Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException]::new()
+                    }
+
+                    $targetResource = Get-TargetResource @getTargetResourceParams
+                    $targetResource.Ensure | Should -Be 'Absent'
+                }
+            }
+
+            Context 'When "Get-ADOrganizationUnit" throws an unexpected error' {
+                It 'Should throw the correct exception' {
+                    Mock -CommandName Get-ADOrganizationalUnit -MockWith { throw 'error' }
+
+                    { Get-TargetResource @getTargetResourceParams } |
+                        Should -Throw ($script:localizedData.GetResourceError -f $getTargetResourceParams.Name)
+                }
             }
         }
         #endregion
 
         #region Function Test-TargetResource
         Describe 'ADOrganizationalUnit\Test-TargetResource' {
-            It 'Returns a "System.Boolean" object type' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $targetResource = Test-TargetResource @testPresentParams
+            BeforeAll {
+                $testTargetResourceParams = @{
+                    Name                            = $mockResource.Name
+                    Path                            = $mockResource.Path
+                    Description                     = $mockResource.Description
+                    ProtectedFromAccidentalDeletion = $mockResource.ProtectedFromAccidentalDeletion
+                }
 
-                $targetResource -is [System.Boolean] | Should -BeTrue
+                $testTargetResourcePresentParams = $testTargetResourceParams.Clone()
+                $testTargetResourcePresentParams['Ensure'] = 'Present'
+
+                $testTargetResourceAbsentParams = $testTargetResourceParams.Clone()
+                $testTargetResourceAbsentParams['Ensure'] = 'Absent'
+
+                Mock -CommandName Assert-Module
             }
 
-            It 'Fails when OU does not exist and "Ensure" = "Present"' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
+            Context 'When the Resource is Present' {
+                BeforeAll {
+                    Mock -CommandName Get-TargetResource -MockWith { $mockGetTargetResourcePresentResult }
+                }
 
-                Test-TargetResource @testPresentParams | Should -BeFalse
+                Context 'When the Resource should be Present' {
+
+                    It 'Should not throw' {
+                        { Test-TargetResource @testTargetResourcePresentParams } | Should -Not -Throw
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $testTargetResourcePresentParams.Name -and `
+                                $Path -eq $testTargetResourcePresentParams.Path } `
+                            -Exactly -times 1
+                    }
+
+                    foreach ($property in $mockChangedResource.Keys)
+                    {
+                        Context "When the $property resource property is not in the desired state" {
+                            BeforeAll {
+                                $testTargetResourceNotInDesiredStateParams = $testTargetResourcePresentParams.Clone()
+                                $testTargetResourceNotInDesiredStateParams.$property = $mockChangedResource.$property
+                            }
+
+                            It 'Should return $false' {
+                                Test-TargetResource @testTargetResourceNotInDesiredStateParams | Should -Be $false
+                            }
+                        }
+                    }
+
+                    Context 'When all the resource properties are in the desired state' {
+                        It 'Should return $true' {
+                            Test-TargetResource @testTargetResourcePresentParams | Should -BeTrue
+                        }
+                    }
+                }
+
+                Context 'When the Resource should be Absent' {
+
+                    It 'Should return $false' {
+                        Test-TargetResource @testTargetResourceAbsentParams | Should -Be $false
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $testTargetResourceAbsentParams.Name -and `
+                                $Path -eq $testTargetResourceAbsentParams.Path } `
+                            -Exactly -times 1
+                    }
+                }
             }
 
-            It 'Fails when OU does exist and "Ensure" = "Absent"' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
+            Context 'When the Resource is Absent' {
+                BeforeAll {
+                    Mock -CommandName Get-TargetResource -MockWith { $mockGetTargetResourceAbsentResult }
+                }
 
-                Test-TargetResource @testAbsentParams | Should -BeFalse
+                Context 'When the Resource should be Present' {
+                    It 'Should return $false' {
+                        Test-TargetResource @testTargetResourcePresentParams | Should -Be $false
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $testTargetResourcePresentParams.Name -and `
+                                $Path -eq $testTargetResourcePresentParams.Path } `
+                            -Exactly -Times 1
+                    }
+                }
+
+                Context 'When the Resource should be Absent' {
+                    It 'Should return $true' {
+                        Test-TargetResource @testTargetResourceAbsentParams | Should -Be $true
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $testTargetResourceAbsentParams.Name -and `
+                                $Path -eq $testTargetResourceAbsentParams.Path } `
+                            -Exactly -times 1
+                    }
+                }
             }
-
-            It 'Fails when OU does exist but "Description" is incorrect' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $testDescriptionParams = $testPresentParams.Clone()
-                $testDescriptionParams['Description'] = 'Wrong description'
-
-                Test-TargetResource @testDescriptionParams | Should -BeFalse
-            }
-
-            It 'Fails when OU does exist but "ProtectedFromAccidentalDeletion" is incorrect' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $testProtectedFromAccidentalDeletionParams = $testPresentParams.Clone()
-                $testProtectedFromAccidentalDeletionParams['ProtectedFromAccidentalDeletion'] = $false
-
-                Test-TargetResource @testProtectedFromAccidentalDeletionParams | Should -BeFalse
-            }
-
-            It 'Passes when OU does exist, "Ensure" = "Present" and all properties are correct' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-
-                Test-TargetResource @testPresentParams | Should -BeTrue
-            }
-
-            It 'Passes when OU does not exist and "Ensure" = "Absent"' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
-
-                Test-TargetResource @testAbsentParams | Should -BeTrue
-            }
-
-            It 'Passes when no OU description is specified with existing OU description' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                $testEmptyDescriptionParams = $testPresentParams.Clone()
-                $testEmptyDescriptionParams['Description'] = ''
-
-                Test-TargetResource @testEmptyDescriptionParams | Should -BeTrue
-            }
-
         }
         #endregion
 
         #region Function Set-TargetResource
+
         Describe 'ADOrganizationalUnit\Set-TargetResource' {
-            It 'Calls "New-ADOrganizationalUnit" when "Ensure" = "Present" and OU does not exist' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
-                Mock -CommandName New-ADOrganizationalUnit
-
-                Set-TargetResource @testPresentParams
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -ParameterFilter `
-                    { $Name -eq $testPresentParams.Name } -Exactly -Times 1 -Scope It
-            }
-
-            It 'Calls "New-ADOrganizationalUnit" when "Ensure" = "Present" and Parent Path does not exist' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
-                Mock -CommandName New-ADOrganizationalUnit -MockWith {
-                    throw New-Object Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException
+            BeforeAll {
+                $setTargetResourceParameters = @{
+                    Name                            = $mockResource.Name
+                    Path                            = $mockResource.Path
+                    Description                     = $mockResource.Description
+                    ProtectedFromAccidentalDeletion = $mockResource.ProtectedFromAccidentalDeletion
                 }
 
-                $errorMessage = $script:localizedData.PathNotFoundError -f $testPresentParams.Path
-                { Set-TargetResource @testPresentParams } | Should -Throw $errorMessage
+                $setTargetResourcePresentParameters = $setTargetResourceParameters.Clone()
+                $setTargetResourcePresentParameters.Ensure = 'Present'
 
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -ParameterFilter `
-                    { $Name -eq $testPresentParams.Name } -Exactly -Times 1 -Scope It
-            }
+                $setTargetResourceAbsentParameters = $setTargetResourceParameters.Clone()
+                $setTargetResourceAbsentParameters.Ensure = 'Absent'
 
-            It 'Calls "New-ADOrganizationalUnit" when "Ensure" = "Present" and an unknown error occurs' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
-                $errorMessage = 'Unknown Error'
-                Mock -CommandName New-ADOrganizationalUnit -MockWith { throw $errorMessage }
-
-                { Set-TargetResource @testPresentParams } | Should -Throw $errorMessage
-
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -ParameterFilter `
-                    { $Name -eq $testPresentParams.Name } -Exactly -Times 1 -Scope It
-            }
-
-            It 'Calls "New-ADOrganizationalUnit" with credentials when specified' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit
                 Mock -CommandName New-ADOrganizationalUnit
-
-                Set-TargetResource @testPresentParams -Credential $testCredential
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -ParameterFilter `
-                    { $Credential -eq $testCredential } -Exactly -Times 1 -Scope It
-            }
-
-            It 'Calls "Set-ADOrganizationalUnit" when "Ensure" = "Present" and OU does exist' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
                 Mock -CommandName Set-ADOrganizationalUnit
-
-                Set-TargetResource @testPresentParams
-                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Scope It
-            }
-
-            It 'Calls "Set-ADOrganizationalUnit" with credentials when specified' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                Mock -CommandName Set-ADOrganizationalUnit -ParameterFilter { $Credential -eq $testCredential }
-
-                Set-TargetResource @testPresentParams -Credential $testCredential
-                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -ParameterFilter { $Credential -eq $testCredential } -Scope It
-            }
-
-            It 'Calls "Remove-ADOrganizationalUnit" when "Ensure" = "Absent" and OU does exist but is unprotected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith {
-                    $unprotectedFakeAdOu = $protectedFakeAdOu.Clone()
-                    $unprotectedFakeAdOu['ProtectedFromAccidentalDeletion'] = $false
-                    return [PSCustomObject] $unprotectedFakeAdOu
-                }
                 Mock -CommandName Remove-ADOrganizationalUnit
-
-                Set-TargetResource @testAbsentParams
-                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Scope It
-            }
-
-            It 'Calls "Remove-ADOrganizationalUnit" when "Ensure" = "Absent" and OU does exist and is protected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                Mock -CommandName Remove-ADOrganizationalUnit
-
-                Set-TargetResource @testAbsentParams
-                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Scope It
-            }
-
-            It 'Calls "Remove-ADOrganizationalUnit" with credentials when specified' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                Mock -CommandName Remove-ADOrganizationalUnit -ParameterFilter { $Credential -eq $testCredential }
-
-                Set-TargetResource @testAbsentParams -Credential $testCredential
-                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -ParameterFilter { $Credential -eq $testCredential } -Scope It
-            }
-
-            It 'Calls "Set-ADOrganizationalUnit" when "Ensure" = "Absent", OU does exist but is protected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-                Mock -CommandName Remove-ADOrganizationalUnit
-                Mock -CommandName Set-ADOrganizationalUnit
-
-                Set-TargetResource @testAbsentParams
-                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Scope It
-            }
-
-            It 'Does not call "Set-ADOrganizationalUnit" when "Ensure" = "Absent", OU does exist but is unprotected' {
-                Mock -CommandName Assert-Module
-                Mock -CommandName Get-ADOrganizationalUnit -MockWith {
-                    $unprotectedFakeAdOu = $protectedFakeAdOu.Clone()
-                    $unprotectedFakeAdOu['ProtectedFromAccidentalDeletion'] = $false
-                    return [PSCustomObject] $unprotectedFakeAdOu
-                }
-                Mock -CommandName Remove-ADOrganizationalUnit
-                Mock -CommandName Set-ADOrganizationalUnit
-
-                Set-TargetResource @testAbsentParams
-                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Scope It -Exactly 0
-            }
-
-            It "Calls Restore-AdCommonObject when RestoreFromRecycleBin is used" {
-                $restoreParam = $testPresentParams.Clone()
-                $restoreParam.RestoreFromRecycleBin = $true
-                Mock -CommandName Get-TargetResource -MockWith { return @{Ensure = 'Absent' } }
-                Mock -CommandName Restore-ADCommonObject -MockWith { return [PSCustomObject] $protectedFakeAdOu }
-
-                Set-TargetResource @restoreParam
-
-                Assert-MockCalled -CommandName Restore-AdCommonObject -Scope It
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Scope It -Exactly -Times 0
-            }
-
-            It "Calls New-ADOrganizationalUnit when RestoreFromRecycleBin is used and if no object was found in the recycle bin" {
-                $restoreParam = $testPresentParams.Clone()
-                $restoreParam.RestoreFromRecycleBin = $true
-                Mock -CommandName Get-TargetResource -MockWith { return @{Ensure = 'Absent' } }
-                Mock -CommandName New-ADOrganizationalUnit
                 Mock -CommandName Restore-ADCommonObject
-
-                Set-TargetResource @restoreParam
-
-                Assert-MockCalled -CommandName Restore-AdCommonObject -Scope It
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Scope It
             }
 
-            It "Throws if the object cannot be restored" {
-                $restoreParam = $testPresentParams.Clone()
-                $restoreParam.RestoreFromRecycleBin = $true
-                Mock -CommandName Get-TargetResource -MockWith { return @{Ensure = 'Absent' } }
-                Mock -CommandName New-ADOrganizationalUnit
-                Mock -CommandName Restore-ADCommonObject -MockWith { throw (New-Object -TypeName System.InvalidOperationException) }
+            Context 'When the Resource is Present' {
+                BeforeAll {
+                    Mock -CommandName Get-TargetResource -MockWith { $mockGetTargetResourcePresentResult }
+                }
 
-                { Set-TargetResource @restoreParam; } | Should -Throw
+                Context 'When the Resource should be Present' {
+                    foreach ($property in $mockChangedResource.Keys)
+                    {
+                        Context "When $property has changed" {
+                            BeforeAll {
+                                $setTargetResourceParametersChangedProperty = $setTargetResourcePresentParameters.Clone()
+                                $setTargetResourceParametersChangedProperty.$property = $mockChangedResource.$property
+                            }
 
-                Assert-MockCalled -CommandName Restore-AdCommonObject -Scope It
-                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Scope It -Exactly -Times 0
+                            It 'Should not throw' {
+                                { Set-TargetResource @setTargetResourceParametersChangedProperty } | Should -Not -Throw
+                            }
+
+                            It 'Should call the correct mocks' {
+                                Assert-MockCalled -CommandName Get-TargetResource `
+                                    -ParameterFilter { $Name -eq $setTargetResourceParametersChangedProperty.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                    -ParameterFilter { $Identity -eq $mockDistinguishedName } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                            }
+                        }
+                    }
+
+                    Context 'When the "Credential" parameter is specified' {
+                        It 'Should not throw' {
+                            { Set-TargetResource @setTargetResourcePresentParameters -Credential $testCredential } |
+                                Should -Not -Throw
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                -ParameterFilter { $Credential -eq $testCredential } -Exactly -Times 1
+                        }
+                    }
+
+                    Context 'When Set-ADOrganizationalUnit throws an exception' {
+                        BeforeAll {
+                            Mock -CommandName Set-ADOrganizationalUnit -MockWith { throw $mockError }
+                        }
+
+                        It 'Should throw the correct exception' {
+                            { Set-TargetResource @setTargetResourcePresentParameters } |
+                                Should -Throw ($script:localizedData.SetResourceError -f
+                                    $setTargetResourcePresentParameters.Name)
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourceAbsentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                -ParameterFilter { $Identity -eq $mockDistinguishedName } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+                    }
+                }
+
+                Context 'When the Resource should be Absent' {
+                    It 'Should not throw' {
+                        { Set-TargetResource @setTargetResourceAbsentParameters } | Should -Not -Throw
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $setTargetResourceAbsentParameters.Name } `
+                            -Exactly -Times 1
+                        Assert-MockCalled -CommandName Remove-ADOrganizationalUnit `
+                            -ParameterFilter { $Identity -eq $mockDistinguishedName } `
+                            -Exactly -Times 1
+                        Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                    }
+
+                    Context 'When the OrganizationalUnit is protected from deletion' {
+                        BeforeAll {
+                            $mockGetTargetResourcePresentProtectedResult = $mockGetTargetResourcePresentResult.Clone()
+                            $mockGetTargetResourcePresentProtectedResult['ProtectedFromAccidentalDeletion'] = $true
+
+                            Mock -CommandName Get-TargetResource -MockWith { $mockGetTargetResourcePresentProtectedResult }
+                        }
+
+                        It 'Should not throw' {
+                            { Set-TargetResource @setTargetResourceAbsentParameters } | Should -Not -Throw
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourceAbsentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                -ParameterFilter { $Identity -eq $mockDistinguishedName -and `
+                                    $ProtectedFromAccidentalDeletion -eq $false } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit `
+                                -ParameterFilter { $Identity -eq $mockDistinguishedName } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+
+                        Context 'When the "Credential" parameter is specified' {
+                            It 'Should not throw' {
+                                { Set-TargetResource @setTargetResourceAbsentParameters -Credential $testCredential } |
+                                    Should -Not -Throw
+                            }
+
+                            It 'Should call the expected mocks' {
+                                Assert-MockCalled -CommandName Get-TargetResource `
+                                    -ParameterFilter { $Name -eq $setTargetResourceAbsentParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                    -ParameterFilter { $Identity -eq $mockDistinguishedName -and `
+                                        $ProtectedFromAccidentalDeletion -eq $false -and `
+                                        $Credential -eq $testCredential } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit `
+                                    -ParameterFilter { $Identity -eq $mockDistinguishedName -and `
+                                        $Credential -eq $testCredential } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                            }
+                        }
+
+                        Context 'When Set-ADOrganizationalUnit throws an exception' {
+                            BeforeAll {
+                                Mock -CommandName Set-ADOrganizationalUnit -MockWith { throw 'error' }
+                            }
+
+                            It 'Should throw the correct exception' {
+                                { Set-TargetResource @setTargetResourceAbsentParameters } | `
+                                        Should -Throw ($script:localizedData.SetResourceError -f
+                                        $setTargetResourceAbsentParameters.Name)
+                            }
+
+                            It 'Should call the expected mocks' {
+                                Assert-MockCalled -CommandName Get-TargetResource `
+                                    -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Set-ADOrganizationalUnit `
+                                    -ParameterFilter { $Identity -eq $mockDistinguishedName -and `
+                                        $ProtectedFromAccidentalDeletion -eq $false } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                            }
+                        }
+                    }
+
+                    Context 'When Remove-ADOrganizationalUnit throws an exception' {
+                        BeforeAll {
+                            Mock -CommandName Remove-ADOrganizationalUnit -MockWith { throw 'error' }
+                        }
+
+                        It 'Should throw the correct exception' {
+                            { Set-TargetResource @setTargetResourceAbsentParameters } | `
+                                    Should -Throw ($script:localizedData.RemoveResourceError -f
+                                    $setTargetResourceAbsentParameters.Name)
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit `
+                                -ParameterFilter { $Identity -eq $mockDistinguishedName } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+                    }
+                }
+            }
+
+            Context 'When the Resource is Absent' {
+                BeforeAll {
+                    Mock -CommandName Get-TargetResource -MockWith { $mockGetTargetResourceAbsentResult }
+                }
+
+                Context 'When the Resource should be Present' {
+                    It 'Should not throw' {
+                        { Set-TargetResource @setTargetResourcePresentParameters } | Should -Not -Throw
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                            -Exactly -Times 1
+                        Assert-MockCalled -CommandName New-ADOrganizationalUnit `
+                            -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                            -Exactly -Times 1
+                        Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                    }
+
+                    Context 'When the "RestoreFromRecycleBin" parameter is specified' {
+                        BeforeAll {
+                            $setTargetResourcePresentRecycleBinParameters = $setTargetResourcePresentParameters.Clone()
+                            $setTargetResourcePresentRecycleBinParameters['RestoreFromRecycleBin'] = $true
+
+                            Mock -CommandName Restore-ADCommonObject -MockWith { $true }
+                        }
+
+                        It 'Should not throw' {
+                            { Set-TargetResource @setTargetResourcePresentRecycleBinParameters } | Should -Not -Throw
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Restore-ADCommonObject `
+                                -ParameterFilter { $Identity -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                        }
+
+                        Context 'When the "Credential" parameter is specified' {
+                            It 'Should not throw' {
+                                { Set-TargetResource @setTargetResourcePresentRecycleBinParameters `
+                                        -Credential $testCredential } | Should -Not -Throw
+                            }
+
+                            It 'Should call the expected mocks' {
+                                Assert-MockCalled -CommandName Get-TargetResource `
+                                    -ParameterFilter { $Name -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Restore-ADCommonObject `
+                                    -ParameterFilter { `
+                                        $Identity -eq $setTargetResourcePresentRecycleBinParameters.Name -and `
+                                        $Credential -eq $testCredential } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            }
+                        }
+
+                        Context 'When Restore from Recycle Bin was unsuccessful' {
+                            BeforeAll {
+                                Mock -CommandName Restore-ADCommonObject -MockWith { $false }
+                            }
+
+                            It 'Should not throw' {
+                                { Set-TargetResource @setTargetResourcePresentRecycleBinParameters } | Should -Not -Throw
+                            }
+
+                            It 'Should call the expected mocks' {
+                                Assert-MockCalled -CommandName Get-TargetResource `
+                                    -ParameterFilter { $Name -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Restore-ADCommonObject `
+                                    -ParameterFilter { $Identity -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName New-ADOrganizationalUnit `
+                                    -ParameterFilter { $Name -eq $setTargetResourcePresentRecycleBinParameters.Name } `
+                                    -Exactly -Times 1
+                                Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                                Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            }
+                        }
+                    }
+
+                    Context 'When the "Credential" parameter is specified' {
+                        It 'Should not throw' {
+                            { Set-TargetResource @setTargetResourcePresentParameters `
+                                    -Credential $testCredential } | Should -Not -Throw
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name -and `
+                                    $Credential -eq $testCredential } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+                    }
+
+                    Context 'When New-ADOrganizationalUnit throws ADIdentityNotFoundException' {
+                        BeforeAll {
+                            $ADIdentityNotFoundException = `
+                                [Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException]::new()
+
+                            Mock -CommandName New-ADOrganizationalUnit -MockWith { throw $ADIdentityNotFoundException }
+                        }
+
+                        It 'Should throw the correct exception' {
+                            { Set-TargetResource @setTargetResourcePresentParameters } |
+                                Should -Throw ($script:localizedData.PathNotFoundError -f
+                                    $setTargetResourcePresentParameters.Path)
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+                    }
+
+                    Context 'When New-ADOrganizationalUnit throws an unexpected exception' {
+                        BeforeAll {
+                            Mock -CommandName New-ADOrganizationalUnit -MockWith { throw $mockError }
+                        }
+
+                        It 'Should throw the correct exception' {
+                            { Set-TargetResource @setTargetResourcePresentParameters } |
+                                Should -Throw ($script:localizedData.NewResourceError -f
+                                    $setTargetResourcePresentParameters.Name)
+                        }
+
+                        It 'Should call the expected mocks' {
+                            Assert-MockCalled -CommandName Get-TargetResource `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName New-ADOrganizationalUnit `
+                                -ParameterFilter { $Name -eq $setTargetResourcePresentParameters.Name } `
+                                -Exactly -Times 1
+                            Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                            Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                        }
+                    }
+                }
+
+                Context 'When the Resource should be Absent' {
+                    It 'Should not throw' {
+                        { Set-TargetResource @setTargetResourceAbsentParameters -Verbose } | Should -Not -Throw
+                    }
+
+                    It 'Should call the expected mocks' {
+                        Assert-MockCalled -CommandName Get-TargetResource `
+                            -ParameterFilter { $Name -eq $setTargetResourceAbsentParameters.Name } `
+                            -Exactly -Times 1
+                        Assert-MockCalled -CommandName New-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Set-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Remove-ADOrganizationalUnit -Exactly -Times 0
+                        Assert-MockCalled -CommandName Restore-ADCommonObject -Exactly -Times 0
+                    }
+                }
             }
         }
         #endregion